requirements.txt
azure-identity==1.15.0
azure-mgmt-compute==30.5.0
azure-mgmt-resource==23.0.1
azure-mgmt-monitor==6.0.2
azure-mgmt-costmanagement==4.0.1
azure-mgmt-security==6.0.0
openai==1.12.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
redis==5.0.1
prometheus-client==0.19.0
pyyaml==6.0.1
aiohttp==3.9.3
structlog==24.1.0